feat: fusion required and non-required activity pages
