+++ agir/activity/components/NotificationSettings/NotificationSettingItem.js
