+++ agir/activity/components/NotificationSettings/NotificationSettingLink.js
