+++ agir/activity/components/NotificationSettings/NotificationSettingPanel.js
