+++ agir/activity/components/NotificationSettings/NotificationSettings.js
