+++ agir/activity/components/common/hooks.js
@@ -1,19 +1,18 @@
-import useSWR from "swr";
-import {
-  activityStatus,
-  getUninteracted,
-  getUnread,
-  setActivityAsInteracted,
-} from "@agir/activity/common/helpers";
 import { useCallback, useMemo } from "react";
+import useSWR from "swr";
+
+import {
+  setActivityAsInteracted,
+  getActivityEndpoint,
+} from "@agir/activity/common/api";
+import { ACTIVITY_STATUS, getUnread } from "@agir/activity/common/helpers";
 
 export const useHasUnreadActivity = () => {
-  const { data: activities } = useSWR("/api/user/activities/", {
+  const { data: activities } = useSWR(getActivityEndpoint("activities"), {
     revalidateOnMount: false,
     revalidateOnFocus: false,
     revalidateOnReconnect: false,
   });
-
   const { data: session } = useSWR(activities ? null : "/api/session/");
 
   return activities
@@ -21,40 +20,18 @@ export const useHasUnreadActivity = () => {
     : session && session.hasUnreadActivities;
 };
 
-export const useRequiredActivityCount = () => {
-  const { data: activities } = useSWR("/api/user/required-activities/", {
-    revalidateOnMount: false,
-    revalidateOnFocus: false,
-    revalidateOnReconnect: false,
-  });
-
-  const { data: session } = useSWR(activities ? null : "/api/session/");
-
-  return activities
-    ? getUninteracted(activities).length
-    : session && session.requiredActionActivitiesCount;
-};
-
 export const useCustomAnnouncement = (slug) => {
   const { data: session } = useSWR("/api/session/");
-  const { data, error, mutate } = useSWR(
-    session?.user && slug ? `/api/user/announcements/custom/${slug}/` : null
+  const { data, mutate, error } = useSWR(
+    session?.user && slug
+      ? getActivityEndpoint("customAnnouncement", { slug })
+      : null
   );
 
-  const announcementId = data?.id;
-  const announcement = useMemo(
-    () => {
-      if (
-        !announcementId ||
-        data?.status === activityStatus.STATUS_INTERACTED
-      ) {
-        return null;
-      }
-      return data;
-    },
-    //eslint-disable-next-line
-    [announcementId]
-  );
+  const announcement =
+    !data?.id || data?.status === ACTIVITY_STATUS.STATUS_INTERACTED
+      ? null
+      : data;
 
   const activityId = announcement?.activityId;
   const dismissCallback = useCallback(async () => {
@@ -66,7 +43,7 @@ export const useCustomAnnouncement = (slug) => {
     mutate(
       (announcement) => ({
         ...announcement,
-        status: activityStatus.STATUS_INTERACTED,
+        status: ACTIVITY_STATUS.STATUS_INTERACTED,
       }),
       false
     );
