--- agir/activity/components/page__activities/activity.config.json
@@ -1,72 +1,0 @@
-{
-  "announcement": {},
-  "waiting-payment": {
-    "icon": "alert-circle"
-  },
-  "group-invitation": {
-    "icon": "mail"
-  },
-  "new-member": {
-    "icon": "user-plus"
-  },
-  "new-message": {
-    "icon": "mail"
-  },
-  "new-comment": {
-    "icon": "mail"
-  },
-  "waiting-location-group": {
-    "icon": "alert-circle"
-  },
-  "group-coorganization-invite": {
-    "icon": "mail"
-  },
-  "waiting-location-event": {
-    "icon": "alert-circle"
-  },
-  "group-coorganization-accepted": {
-    "icon": "calendar",
-    "hasEvent": true
-  },
-  "group-info-update": {
-    "icon": "info"
-  },
-  "accepted-invitation-member": {
-    "icon": "user-plus"
-  },
-  "new-attendee": {
-    "icon": "user-plus"
-  },
-  "event-update": {
-    "icon": "info",
-    "hasEvent": true
-  },
-  "new-event-mygroups": {
-    "icon": "calendar",
-    "hasEvent": true
-  },
-  "new-report": {
-    "icon": "file-text",
-    "hasEvent": true
-  },
-  "event-suggestion": {
-    "icon": "calendar",
-    "hasEvent": true
-  },
-  "group-coorganization-info": {
-    "icon": "calendar",
-    "hasEvent": true
-  },
-  "cancelled-event": {
-    "icon": "x-circle"
-  },
-  "referral-accepted": {
-    "icon": "share-2"
-  },
-  "transferred-group-member": {
-    "icon": "info"
-  },
-  "new-members-through-transfer": {
-    "icon": "user-plus"
-  }
-}
