+++ agir/front/components/dashboardComponents/navigation.config.json
@@ -18,15 +18,6 @@
       "secondaryLinks": "groups__personGroups"
     },
     {
-      "id": "requiredActivities",
-      "icon": "bell",
-      "title": "À faire",
-      "to": "requiredActivities",
-      "requiredActivityCounter": true,
-      "desktop": true,
-      "mobile": true
-    },
-    {
       "id": "activities",
       "icon": "align-left",
       "title": "Actualités",
