+++ agir/front/components/navigationPage/NavigationPage.js
@@ -15,7 +15,6 @@ import Link from "@agir/front/app/Link";
 
 import { routeConfig } from "@agir/front/app/routes.config";
 import CONFIG from "@agir/front/dashboardComponents/navigation.config";
-import { useRequiredActivityCount } from "@agir/activity/common/hooks";
 
 const MAIN_LINKS = CONFIG.menuLinks.filter(({ mobile }) => mobile === false);
 
@@ -108,7 +107,6 @@ MenuLink.propTypes = {
 };
 
 const NavigationPage = ({ active }) => {
-  const requiredActionActivityCount = useRequiredActivityCount();
   const routes = useSelector(getRoutes);
   return (
     <Navigation>
@@ -128,7 +126,6 @@ const NavigationPage = ({ active }) => {
                   ? routeConfig[link.to].path
                   : undefined
               }
-              counter={link.counter && requiredActionActivityCount}
             />
           ) : null
         )}
