+++ agir/front/components/serviceWorker/cachedRoutes.config.js
@@ -4,7 +4,6 @@ const config = [
   "/api/evenements/suggestions/",
   "/api/groupes/",
   "/api/user/activities/",
-  "/api/user/required-activities/",
   new RegExp(".+/api/evenements/[0-9a-f-]+/$"),
 ];
